core: move subscription types into their own module
